Clarify the caveats on the Next.js middleware docs
The wording of the previous caveat section was potentially misleading. Previously, the second had two points, starting with:

- `Currently only supports session verification...`
- and `Only supports the "jwt" session strategy...`

As someone using database-backed sessions, this wording was very confusing — the first caveat told me that sessions would work, but the second part told me that only `jwt` sessions work.

I think that simplifying this down to a single caveat, that explains that only the `jwt` session strategy is supported will make this more clear that the middleware option is not valid for anyone using database-backed sessions.